% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Load base gridded hexagonal map for metropolitan France}
\usage{
get_map()
}
\value{
a sf object with gridded hexagonal map for metropolitan France
}
\description{
Function to add the gridded hexagonal map of metropolitan France.
This function has no argument
}
\examples{
# Plot the map
library(sf)
library(tidyverse)

get_map()\%>\%
  ggplot()+
  geom_sf()
}
