% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_layer.R
\name{get_time_layer}
\alias{get_time_layer}
\title{Add time series layers}
\usage{
get_time_layer(layer, from, to)
}
\arguments{
\item{layer}{type of data to be added ("temperature" for monthly mean temperature in °C or "rainfall" for sum of monthly rainfall in mm). Climatic data layers (temperature and rainfall) are available from 2000 to 2020.}

\item{from}{starting year (from 2000 to 2020)}

\item{to}{ending year (may be the same as starting year to extract only one year of data)}
}
\value{
a tibble with the hexagon number ("hex_id" column) and the selected features values
}
\description{
Add time series data layer, based on the spatial grid of the map obtained by the get_map() function
}
\examples{
# Get mean temperature from 2000 to 2020 and filter for May only
library(sf)
library(tidyverse)

data<-get_time_layer("temperature",from=2000, to=2020)\%>\%
  filter(month==5)

}
\references{
Climate data from \href{https://developers.google.com/earth-engine/datasets/catalog/ECMWF_ERA5_MONTHLY}{ERA5} (Copernicus Climate Change Service)
}
