% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map.R
\name{get_map}
\alias{get_map}
\title{Load}
\usage{
get_map()
}
\value{
gridded hexagonal map of metropolitan France
}
\description{
Function to add the gridded hexagonal map of metropolitan France.
This function has no argument
}
\examples{
# Load map
get_map()\%>\%
ggplot()+
geom_sf()
}
