% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_static_layer.R
\name{get_static_layer}
\alias{get_static_layer}
\title{Add static data layers}
\usage{
get_static_layer(layer)
}
\arguments{
\item{layer}{type of data to be added ("soil" for soil feature, "crops" for main crops distribution, "herds" for farm animal density)}
}
\value{
a tibble with the hexagon number ("hex_id" column) and the selected features values
}
\description{
Add static data layer (not updated from year to year), based on the spatial grid of the map obtained by the get_map() function
}
\examples{
# Add soil features to grid then plot clay content
library(sf)
library(tidyverse)

get_map()\%>\%
  left_join(get_static_layer("soil"))\%>\%
  ggplot()+
  geom_sf(aes(fill=clay_0to30cm_percent)) 

}
